detection tree gives hieght width + many other fixes
