--- cusub_common/motor_controllers/special_conf/sim_elev2depth.yaml
@@ -1,4 +1,0 @@
-Kp: -1.5
-Kd: -5.0
-Ki: -2.0
-windup_limit: 10
