--- cusub_common/motor_controllers/special_conf/sim_elevAz2drive.yaml
@@ -1,4 +1,0 @@
-Kd: 0.0
-Ki: 0.0
-Kp: -0.5
-windup_limit: 10
