--- cusub_common/motor_controllers/special_conf/sim_mag2drive.yaml
@@ -1,4 +1,0 @@
-Kd: 0.0005
-Ki: 0.0001
-Kp: 0.0002
-windup_limit: 10
