+++ cusub_cortex/state_machine/src/tasks/jiangshi.py
@@ -1,250 +1,291 @@
 #!/usr/bin/env python
-from __future__ import division
+# -*- coding: utf-8 -*-
+from __future__ import division, print_function
 """
 Jiangshi Buoy Task, attempts to bump into the jiangshi buoy
 Objectives:
 - Search
+- Approach
 - Slay
----> Go to approach point
----> Slay vampire
----> Backup
+- Back Up
 """
-from tasks.task import Task, Objective
+from tasks.task import Task, Objective, Timeout
 from tasks.search import Search
-import tf
-import numpy as np
+from tasks.pid_client import PIDClient
 import rospy
-from geometry_msgs.msg import PoseStamped, Pose
+import tf
 import smach
 import smach_ros
-import copy
-from perception_control.msg import VisualServoAction, VisualServoGoal, VisualServoFeedback
-import actionlib
-from std_msgs.msg import Float64
-from waypoint_navigator.srv import *
+from geometry_msgs.msg import Pose, Point
+from std_msgs.msg import Time
+from detection_listener.listener import DetectionListener
+from waypoint_navigator.srv import ToggleControl
+import numpy as np
+from cusub_print.cuprint import bcolors
+import sys
 
 class Jiangshi(Task):
-    name = "jiangshi"
+    name = "Jiangshi"
 
     def __init__(self):
-        super(Jiangshi, self).__init__()
+        super(Jiangshi, self).__init__(self.name)
+
+        # All Objectives share the same listener to gaurantee same data between objectives
+        self.listener = DetectionListener() 
+
         self.init_objectives()
         self.link_objectives()
 
     def init_objectives(self):
-        search_frames = ["leviathan/description/occam0_frame_optical"]
         search_classes = ["vampire_cute"]
-        self.search = Search.from_bounding_box(self.get_prior_param(), search_classes, search_frames)
-        self.slay = Slay()
+        self.search = Search(self.name, self.listener, search_classes, self.get_prior_param())
+        self.approach = Approach(self.name, self.listener)
+        self.slay = Slay(self.name)
+        self.retrace = Retrace(self.name, self.listener)
+        self.backup = Backup(self.name)
 
     def link_objectives(self):
         with self:
-            smach.StateMachine.add('Search', self.search, transitions={'found':'Slay', 'not_found':'manager'}, \
+            smach.StateMachine.add('Search', self.search, transitions={'found':'Approach', 'not_found':'manager'}, \
                 remapping={'timeout_obj':'timeout_obj', 'outcome':'outcome'})
-            smach.StateMachine.add('Slay', self.slay, transitions={'success':'manager', 'timed_out':'manager'}, \
+            smach.StateMachine.add('Retrace', self.retrace, transitions={'found':'Approach', 'not_found':'Search'}, \
                 remapping={'timeout_obj':'timeout_obj', 'outcome':'outcome'})
+            smach.StateMachine.add('Slay', self.slay, transitions={'slayed':'Backup', 'timed_out':'manager'}, \
+                remapping={'timeout_obj':'timeout_obj', 'outcome':'outcome'})
+            smach.StateMachine.add('Backup', self.backup, transitions={'backed_up':'manager', 'timed_out':'manager'}, \
+                remapping={'timeout_obj':'timeout_obj', 'outcome':'outcome'})
+            smach.StateMachine.add('Approach', self.approach, transitions={'in_position':'Slay', 'timed_out':'manager', 'lost_object':'Retrace'}, \
+                remapping={'timeout_obj':'timeout_obj', 'outcome':'outcome'})
+            
+
+class Approach(Objective):
+    outcomes = ['in_position','timed_out', 'lost_object']
+
+    target_class_id = "vampire_cute"
+
+    def __init__(self, task_name, listener):
+        name = task_name + "/Approach"
+        super(Approach, self).__init__(self.outcomes, name)
+        self.listener = listener
+        self.yaw_client = PIDClient(name, "yaw")
+        self.drive_client = PIDClient(name, "drive")
+        self.depth_client = PIDClient(name, "depth")
+
+        self.retrace_timeout = rospy.get_param("tasks/" + task_name.lower() + "/retrace_timeout", 2)
+        seconds = rospy.get_param("tasks/" + task_name.lower() + "/seconds_in_position", 2)
+        self.rate = rospy.get_param("tasks/" + task_name.lower() + "/new_dv_check_rate", 30)
+        self.count_target = seconds * self.rate
+        self.count = 0
+        
+        self.approach_stop_dist = rospy.get_param("tasks/" + task_name.lower() + "/approach_stop_dist")
+        self.approach_carrot_dist = rospy.get_param("tasks/" + task_name.lower() + "/approach_carrot_dist")
+        self.height2dist_slope = rospy.get_param("tasks/" + task_name.lower() + "/height2dist_slope")
+        self.height2dist_intercept = rospy.get_param("tasks/" + task_name.lower() + "/height2dist_intercept")
+    
+    def execute(self, userdata):
+        self.cuprint("executing")
+        self.toggle_waypoint_control(True)
+
+        # Find vampire_cute's dobject number and check for errors
+        dobj_nums = self.listener.query_class(self.target_class_id)
+        if len(dobj_nums) > 1: # Check if more than 1 instance of target_class
+            self.cuprint(str(len(dobj_nums)) + " " + self.target_class_id + " classes detected!", warn=True)
+            self.cuprint("selecting the first", warn=True)
+        elif not dobj_nums: # Chck if target class is not present (shouldn't be possible)
+            self.cuprint("somehow no " + self.target_class_id + " classes found?", warn=True)
+            return "lost_object"
+        dobj_num = dobj_nums[0]
+        self.cuprint("located " + self.target_class_id + "'s dobject num: " + bcolors.HEADER + str(dobj_num) + bcolors.ENDC)
+        
+        # Start a watch dog timer on detections
+        watchdog_timer = Timeout(self.name + "/Retrace Watchdog")
+
+        # Enable the PID loops
+        self.yaw_client.enable()
+        self.drive_client.enable()
+        self.depth_client.enable()
+
+        self.cuprint("servoing")
+        r = rospy.Rate(self.rate)
+        print("")
+        while not rospy.is_shutdown():
+            if self.listener.check_new_dv(dobj_num):
+                watchdog_timer.set_new_time(self.retrace_timeout, print_new_time=False)
+                [az, el, height, width] = self.listener.get_avg_bearing(dobj_num, num_dv=3)
+                if az == None: # if secs is set in get_avg_bearing we could have a new dv it just wasn't in recent time frame
+                    r.sleep()
+                    continue
+                self.yaw_client.set_setpoint(az, loop=False)
+
+                # Calculate translational errors
+                dist = self.get_jiangshi_distance(height)
+                [delta_x, delta_y, delta_z] = self.get_relative_jiangshi_xyz(az, el, dist)
+                xy_dist = np.linalg.norm([delta_x, delta_y])
+
+                # set depth setpoint
+                self.depth_client.set_setpoint(self.depth_client.get_standard_state() + delta_z, loop=False)
+                
+                # pretty printing
+                error_string_dist = "Error xy: " + bcolors.HEADER + str(round(xy_dist,2)) +bcolors.ENDC
+                error_string_z = " | z: " + bcolors.HEADER + str(round(delta_z,2)) + bcolors.ENDC
+                self.cuprint(error_string_dist + error_string_z, print_prev_line=True)
+
+                # Set drive setpoint and check if in position
+                if xy_dist > self.approach_stop_dist:
+                    self.drive_client.set_setpoint(self.drive_client.get_standard_state() + self.approach_carrot_dist, loop=False)
+                else:
+                    self.drive_client.set_setpoint(self.drive_client.get_standard_state())
+                    break
+
+            elif watchdog_timer.timed_out:
+                self.yaw_client.disable()
+                self.drive_client.disable()
+                self.depth_client.disable()
+                return "lost_object"
+
+            if userdata.timeout_obj.timed_out:
+                self.yaw_client.disable()
+                self.drive_client.disable()
+                self.depth_client.disable()
+                userdata.outcome = "timed_out"
+                return "not_found"
+            r.sleep()
+
+        watchdog_timer.timer.shutdown()
+        self.yaw_client.disable()
+        self.drive_client.disable()
+        self.depth_client.disable()
+        return "in_position"
+
+    def get_relative_jiangshi_xyz(self, az, el, dist):
+        # Converts spherical coordinates to x,y,z
+        z = dist * np.sin(-el) # flip elevation sign
+        xy_dist = dist * np.cos(el)
+        x = xy_dist * np.cos(az)
+        y = xy_dist * np.sin(az)
+        return [x,y,z]
+    
+    def get_jiangshi_distance(self, jiangshi_height):
+        # Solve linear system
+        return self.height2dist_slope * jiangshi_height + self.height2dist_intercept
 
 class Slay(Objective):
-    """
-    Go to a point in front of the bouy, slay jiangshi backwards, backup
-    """
-    outcomes=['success', 'timed_out']
+    outcomes = ['slayed', 'timed_out']
 
-    def __init__(self):
-        super(Slay, self).__init__(self.outcomes, "Slay")
-        self.jiangshi_pose = None
-        rospy.Subscriber("cusub_cortex/mapper_out/jiangshi", PoseStamped, self.jiangshi_callback)
-        self.approach_dist = rospy.get_param('tasks/jiangshi/solvepnp/approach_dist', 2.0)
-        self.print_configuration()
-
-        # SOLVEPNP
-        self.replan_threshold = rospy.get_param('tasks/jiangshi/solvepnp/replan_threshold', 0.5)
-        self.use_buoys_yaw = rospy.get_param('tasks/jiangshi/solvepnp/use_buoys_yaw', True)
-        self.slay_dist = rospy.get_param('tasks/jiangshi/solvepnp/slay_dist', 0.5) 
-
-        # VISUAL SERVO
-        self.feedback = None
-        self.target_pixel_box = rospy.get_param("tasks/jiangshi/visual_servo/target_pixel_threshold")
-        self.vs_client = actionlib.SimpleActionClient('visual_servo', VisualServoAction)
-        rospy.loginfo("...waiting for visual_servo server")
-        self.vs_client.wait_for_server()
-        rospy.loginfo("\tfound visual servo server")
-
-        self.drive_state = None
-        rospy.Subscriber("cusub_common/motor_controllers/pid/drive/state", Float64, self.drive_callback)
-        self.drive_pub = rospy.Publisher("cusub_common/motor_controllers/pid/drive/setpoint", Float64, queue_size=1)
-
-        rospy.loginfo("...waiting for toggle waypoint control")
-        rospy.wait_for_service('cusub_common/toggleWaypointControl')
-        self.wayToggle = rospy.ServiceProxy('cusub_common/toggleWaypointControl', ToggleControl)
-        rospy.loginfo("\tfound toggle waypoint control")
-
-    def drive_callback(self, msg):
-        self.drive_state = msg.data
-
-    def print_configuration(self):
-        if rospy.get_param("tasks/jiangshi/method") == "solvepnp":
-            rospy.loginfo("...method: SOLVEPNP")
-        else:
-            rospy.loginfo("...method: VISUAL_SERVO")
-
-    def jiangshi_callback(self, msg):
-        # Set the first pose and don't replan
-        if self.jiangshi_pose == None:
-            self.jiangshi_pose = msg
-            return
-        change_in_pose = self.get_distance_xy(self.jiangshi_pose.pose.position, msg.pose.position)
-        if (change_in_pose > self.replan_threshold) and not self.replan_requested():
-            self.jiangshi_pose = msg
-            self.request_replan()
-
-    def vs_feedback_callback(self, feedback):
-        self.feedback = feedback.centered
-    
-    def get_approach_pose(self):
-        approach_pose = Pose()
-        if self.use_buoys_yaw:    
-            quat = [self.jiangshi_pose.pose.orientation.x, self.jiangshi_pose.pose.orientation.y,self.jiangshi_pose.pose.orientation.z,self.jiangshi_pose.pose.orientation.w]
-            jiangshi_roll, jiangshi_pitch, jiangshi_yaw = tf.transformations.euler_from_quaternion(quat)
-            approach_pose.position.x = self.jiangshi_pose.pose.position.x + self.approach_dist * np.cos(jiangshi_yaw)
-            approach_pose.position.y = self.jiangshi_pose.pose.position.y + self.approach_dist * np.sin(jiangshi_yaw)
-            approach_pose.position.z = self.jiangshi_pose.pose.position.z
-            goal_quat = tf.transformations.quaternion_from_euler(jiangshi_roll, jiangshi_pitch, jiangshi_yaw)
-            approach_pose.orientation.x = goal_quat[0]
-            approach_pose.orientation.y = goal_quat[1]
-            approach_pose.orientation.z = goal_quat[2]
-            approach_pose.orientation.w = goal_quat[3]
-        else: # just draw a line from sub to buoy and hit it
-            approach_pose = self.get_pose_between(self.cur_pose, self.jiangshi_pose.pose, self.approach_dist) # inherited
-        return approach_pose
-
-    def get_slay_pose(self):
-        """
-        Calculate slay pose for the buoy setpoint method
-
-        Parameters
-        ----------
-        self.jiangshi_pose : PoseStamped
-            Jiangshi's pose
-        self.use_buoys_yaw : bool
-            Whether to calculate slay path and approach pose based on the buoys yaw
-        self.slay_dist : float
-            Meters to go behind jiangshi in order to slay him
-
-        Returns
-        -------
-        slay_pose : Pose
-        """
-        slay_pose = Pose()
-        if self.use_buoys_yaw:    
-            quat = [self.jiangshi_pose.pose.orientation.x, self.jiangshi_pose.pose.orientation.y,self.jiangshi_pose.pose.orientation.z,self.jiangshi_pose.pose.orientation.w]
-            jiangshi_roll, jiangshi_pitch, jiangshi_yaw = tf.transformations.euler_from_quaternion(quat)
-            slay_pose.position.x = self.jiangshi_pose.pose.position.x - self.slay_dist * np.cos(jiangshi_yaw)
-            slay_pose.position.y = self.jiangshi_pose.pose.position.y - self.slay_dist * np.sin(jiangshi_yaw)
-            slay_pose.position.z = self.jiangshi_pose.pose.position.z
-        else: # just draw a line from sub to buoy and hit it
-            slay_pose = self.get_pose_behind(self.cur_pose, self.jiangshi_pose.pose, self.slay_dist)
-        return slay_pose
-    
-    def solvepnp_method(self, userdata):
-        rospy.loginfo("...using solvepnp")
-        self.configure_darknet_cameras([1,1,0,0,1,0])
-
-        # Approach 
-        approach_pose = self.get_approach_pose() 
-        while not rospy.is_shutdown():          # Loop until we find a good task pose
-            approach_pose = self.get_approach_pose() #only use the approach part
-            if self.go_to_pose(approach_pose, userdata.timeout_obj):
-                if userdata.timeout_obj.timed_out:
-                    userdata.outcome = "timed_out"
-                    return "timed_out"
-                else: # Loop, replan_requested() happened
-                    pass
-            else: # we've reached our pose!
-                break
-        
-        # Slay
-        if rospy.get_param("tasks/jiangshi/visual_servo/slay_timeout"):
-            userdata.timeout_obj.set_new_time(rospy.get_param("tasks/jiangshi/visual_servo/slay_timeout"))
-        rospy.loginfo("...slaying jiangshi")
-        slay_pose = self.get_slay_pose()
-        self.go_to_pose(slay_pose, userdata.timeout_obj)
-        rospy.loginfo("\tslayed")
-
-        # backup to our previous pose
-        userdata.timeout_obj.set_new_time(0)
-        userdata.timeout_obj.timed_out = False
-        rospy.loginfo("...disabling timeouts while backing up from slay")
-        self.go_to_pose(approach_pose, userdata.timeout_obj, replan_enabled=False, move_mode="backup")
-        userdata.outcome = "success"
-        return "success"
-
-    def visual_servo_method(self, userdata):
-        self.configure_darknet_cameras([1,0,0,0,0,0])
-        rospy.loginfo("...using visual servoing approach")
-        goal = VisualServoGoal()
-        goal.target_classes = ["vampire_cute"]
-        goal.camera = goal.OCCAM
-        goal.x_axis = goal.YAW_AXIS
-        goal.y_axis = goal.DEPTH_AXIS
-        goal.area_axis = goal.DRIVE_AXIS
-        goal.target_frame = rospy.get_param("~robotname") +"/description/occam0_frame_optical"
-        goal.visual_servo_type = goal.PROPORTIONAL
-        goal.target_pixel_x = goal.CAMERAS_CENTER_X
-        goal.target_pixel_y = goal.CAMERAS_CENTER_Y
-        goal.target_box_area = (int) ( goal.ONE_HUNDRED_PERCENT_IMAGE * rospy.get_param("tasks/jiangshi/visual_servo/goal_percentage_of_frame") )
-        goal.target_pixel_threshold = self.target_pixel_box
-        rospy.loginfo("...centering")
-        self.vs_client.send_goal(goal, feedback_cb=self.vs_feedback_callback)
-
-        while not rospy.is_shutdown() :
-            if userdata.timeout_obj.timed_out:
-                self.vs_client.cancel_goal()
-                userdata.outcome = "timed_out"
-                return "timed_out"
-            if self.feedback:
-                break
-            rospy.sleep(0.25)
-        self.vs_client.cancel_goal()
-        rospy.loginfo("\tcentered")
-
-        # Slay the buoy
-        rospy.sleep(2) # Wait for the vs client to release control to the waypoint nav for us to take it back again
-        rospy.loginfo("...slaying")
-        try:
-            res = self.wayToggle(False)
-            rospy.loginfo(res)
-        except rospy.ServiceException, e:
-            rospy.logwarn("Service call failed: %s", e)
-        slay_set = Float64()
-        original_set = copy.deepcopy(self.drive_state)
-        userdata.timeout_obj.set_new_time(rospy.get_param("tasks/jiangshi/visual_servo/slay_timeout"))
-        slay_carrot = rospy.get_param("tasks/jiangshi/visual_servo/slay_carrot")
-        while not rospy.is_shutdown():
-            slay_set.data = self.drive_state + slay_carrot
-            self.drive_pub.publish(slay_set)
-            if userdata.timeout_obj.timed_out:
-                break
-            rospy.sleep(0.25)
-        self.update_next_priors("jiangshi")
-        rospy.loginfo("...slayed, backing up")
-        userdata.timeout_obj.set_new_time(rospy.get_param("tasks/jiangshi/visual_servo/slay_timeout"))
-        # slay_set.data = original_set
-        while not rospy.is_shutdown():
-            slay_set.data = self.drive_state - slay_carrot
-            self.drive_pub.publish(slay_set)
-            if userdata.timeout_obj.timed_out:
-                break
-            rospy.sleep(0.25)
-        self.wayToggle(True)
-
-        userdata.outcome = "success"
-        return "success"
+    def __init__(self, task_name):
+        name = task_name + "/Slay"
+        super(Slay, self).__init__(self.outcomes, name)
+        self.drive_client = PIDClient(name, "drive")
+        self.slay_carrot = rospy.get_param("tasks/" + task_name.lower() + "/slay_carrot")
+        self.approach_stop_dist = rospy.get_param("tasks/" + task_name.lower() + "/approach_stop_dist")
 
     def execute(self, userdata):
-        if rospy.get_param("tasks/jiangshi/method") == "vs": # Do visual servo
-            rospy.loginfo("...enacting visual servoing")
-            return self.visual_servo_method(userdata)
-        else:
-            rospy.loginfo("...using solvepnp")
-            return self.solvepnp_method(userdata)
+        self.cuprint("executing")
+        self.toggle_waypoint_control(True)
+        self.drive_client.enable()
+        cur_state = self.drive_client.get_standard_state()
+        self.drive_client.set_setpoint(cur_state + self.slay_carrot + self.approach_stop_dist)
+        rospy.sleep(5)
+        self.cuprint("slayed")
+        self.drive_client.disable()
+        return "slayed"
+
+class Backup(Objective):
+    outcomes = ['backed_up','timed_out']
+
+    def __init__(self, task_name):
+        name = task_name + "/Backup"
+        super(Backup, self).__init__(self.outcomes, name)
+        self.drive_client = PIDClient(name, "drive")
+        self.carrot_dist = rospy.get_param("tasks/" + task_name.lower() + "/approach_stop_dist")
+
+    def execute(self, userdata):
+        self.cuprint("executing")
+        cur_state = self.drive_client.get_standard_state()
+        self.drive_client.enable()
+        self.drive_client.set_setpoint(cur_state - self.carrot_dist + 0.3) # add a little more so we hit the buoy
+        rospy.sleep(5)
+        self.toggle_waypoint_control(False)
+        self.drive_client.disable()
+        self.cuprint("backed up")
+        return "backed_up"
+
+
+class Retrace(Objective):
+    outcomes = ['found','not_found']
+
+    target_class_id = "vampire_cute"
+
+    def __init__(self, task_name, listener):
+        super(Retrace, self).__init__(self.outcomes, task_name + u"/Rétrace".encode("utf-8"))
+        self.listener = listener
+        self.retrace_hit_cnt = rospy.get_param("tasks/" + task_name.lower() + "/retrace_hit_count")
+        self.retrace_back_in_time = rospy.get_param("tasks/" + task_name.lower() + "/retrace_back_in_time")
+
+    def execute(self, userdata):
+        self.configure_darknet_cameras([1,1,1,1,1,0])
+        self.toggle_waypoint_control(False)
+        self.cuprint("executing")
+        dobj_nums = self.listener.query_class(self.target_class_id)
+        if not len(dobj_nums): #shouldn't be possible
+            # do some timeout?
+            userdata.outcome = "timed_out"
+            return "not_found"
+
+        #This is an assumption that we only have one dobj, which we *should*
+        dobj_num = dobj_nums[0]
+
+        # loop variables
+        count = 0
+        retraced_steps = 1
+        first_bc = self.listener[dobj_num][-1]
+
+        backup_time = first_bc.camera_header.stamp - rospy.Duration(self.retrace_back_in_time)
+        
+        breadcrumb_dvs = self.listener[dobj_num].get_dvectors_since_time(backup_time)
+
+        len_dvec = len(breadcrumb_dvs)
+        self.cuprint("breadcrumbs to follow since " + str(self.retrace_back_in_time) + " seconds ago:" + str(len_dvec))
+        last_sub_pt = breadcrumb_dvs[len_dvec-retraced_steps].sub_pt 
+        last_pose = Pose(last_sub_pt, self.cur_pose.orientation)
+
+        # Start Retrace: Set first waypoint
+        self.go_to_pose_non_blocking(last_pose, move_mode="strafe")
+        # necessary for same-line printing
+        print("")
+        while not rospy.is_shutdown():
+            if self.listener.check_new_dv(dobj_num) and count < self.retrace_hit_cnt: 
+                #found object again
+                count += 1
+            elif count >= self.retrace_hit_cnt:
+                self.cancel_way_client_goal()
+                return "found"
+            else:
+                dist = self.get_distance(self.cur_pose.position, last_pose.position)
+                self.cuprint("Breadcrumb #"+ bcolors.HEADER + str( retraced_steps)+ bcolors.ENDC+ " distance: " + bcolors.OKBLUE + str(dist) + bcolors.ENDC, print_prev_line=True)
+                #self.cuprint("distance to goal: "+ str())
+                if self.check_reached_pose(last_pose, 0.7):
+                    self.cancel_way_client_goal()
+                    retraced_steps += 1
+                    if (retraced_steps > len_dvec):
+                        #means we retraces all our steps. We're lost.
+                        return "not_found"
+                    # Goto Last dvector
+                    # get last dvector sub_pt
+                    last_sub_pt = breadcrumb_dvs[len_dvec-retraced_steps].sub_pt 
+            
+                    last_pose = Pose(last_sub_pt, self.cur_pose.orientation)
+                    #set waypoint to this point. Give some distance to account for error in bearing and sub_pt
+                    # Set waypoint
+                    self.go_to_pose_non_blocking(last_pose, move_mode="strafe", log_print=False)
+
+            if userdata.timeout_obj.timed_out:
+                self.cancel_way_client_goal()
+                userdata.outcome = "timed_out"
+                return "not_found"
+
+        #clean up if we are killed
+        return "not_found"
+        
+
+
