+++ cusub_cortex/state_machine/src/tasks/startup_task.py
@@ -12,15 +12,15 @@ from std_msgs.msg import Float64
 from tasks.task import Task, Objective
 
 class Startup(Task):
-    name = "startup"
+    name = "Startup"
 
     def __init__(self):
-        super(Startup, self).__init__() # become a state machine first
+        super(Startup, self).__init__(self.name) # become a state machine first
         self.init_objectives()
         self.link_objectives()
 
     def init_objectives(self):
-        self.dive = Dive()
+        self.dive = Dive(self.name)
 
     def link_objectives(self):
         with self: # we are a StateMachine
@@ -31,21 +31,22 @@ class Dive(Objective):
 
     outcomes=['done']
 
-    def __init__(self):
-        super(Dive, self).__init__(self.outcomes, "Dive")
+    def __init__(self, task_name):
+        name = task_name + "/Dive"
+        super(Dive, self).__init__(self.outcomes, name)
         self.dive_pub = rospy.Publisher("cusub_common/motor_controllers/pid/depth/setpoint", Float64, queue_size=1)
 
     def execute(self, userdata):
         depth_msg = Float64()
         depth_msg.data = 0.2 # experimentally determined from leviathan
-        rospy.loginfo("...pull tether")
+        self.cuprint("PULL TETHER")
         while not rospy.is_shutdown():
             self.dive_pub.publish(depth_msg)
             if userdata.timeout_obj.timed_out:
                 break
             rospy.sleep(0.25)
         depth_msg.data = rospy.get_param("tasks/startup/depth")
-        rospy.loginfo("...diving")
+        self.cuprint("diving")
         userdata.timeout_obj.set_new_time(rospy.get_param("tasks/startup/dive_time"))
         while not rospy.is_shutdown():
             self.dive_pub.publish(depth_msg)
