+++ cusub_perception/detection_tree/msg/Dvector.msg
@@ -3,7 +3,8 @@ std_msgs/Header camera_header
 geometry_msgs/Point sub_pt
 float64 azimuth
 float64 elevation
-int64 magnitude # number of box pixels
+int64 box_height
+int64 box_width
 string class_id
 int32 dobject_num
 float64 probability
